(actually) fix .gitignore, bug fixes + colors

+++ requirements.sh
@@ -1,63 +1,75 @@
 #!/bin/bash
-error_exit()
-{
-    echo "Error: $1"
-    exit 1
+
+cecho(){
+    RED="\033[0;31m"
+    GREEN="\033[0;32m"  # <-- [0 means not bold
+    YELLOW="\033[0;33m" # <-- [1 means bold
+    CYAN="\033[0;36m"
+    # ... Add more colors if you like
+
+    NC="\033[0m" # No Color
+
+    # printf "${(P)1}${2} ${NC}\n" # <-- zsh
+    printf "${!1}${2} ${NC}\n" # <-- bash
 }
 
+error_exit()
+{
+    cecho "RED" "Error: $1"
+    exit 1
+}
 
 
 macOSversion=$(sw_vers | head -n2 | tail -n1 | cut -c 17-)
 verscheck=$(bc <<<"${macOSversion} < 10.14")
 
 if [ "$(uname)" = "Darwin" ]; then
-  echo "[!] macOS detected!"
+  cecho "CYAN" "[!] macOS detected!"
   if [[ $(sysctl -n machdep.cpu.brand_string) =~ "Apple" ]]; then
-    echo "[!] Apple Silicon detected"
+    cecho "CYAN" "[!] Apple Silicon detected"
     OS="macOS-arm64"
   elif [[ $(sysctl -n machdep.cpu.brand_string) =~ "Intel" ]]; then
-    echo [!] "Intel mac detected!"
+    cecho "CYAN" "[!] Intel mac detected!"
     OS="macOS-x86_64"
   fi
 else
-  echo "Not running on macOS... exiting..."
+  cecho "RED" "Not running on macOS... exiting..."
   exit 2
 fi
 
 if command -v brew >/dev/null && [ -f /opt/procursus/bin/apt ]; then
-  echo "[!] Homebrew and Procursus were found."
+  cecho "GREEN" "[!] Homebrew and Procursus were found."
   echo "Choose a package manager for your dependencies."
   PS3='Please enter your choice: '
   options=("Procursus (apt)" "Homebrew" "Quit")
   select opt in "${options[@]}"; do
     case $opt in
     "Procursus (apt)")
-      echo "[!] Procursus selected."
+      cecho "CYAN" "[!] Procursus selected."
       pkg="sudo apt"
       break
       ;;
     "Homebrew")
-      echo "[!] Homebrew selected."
+      cecho "CYAN" "[!] Homebrew selected."
       pkg="brew"
       break
       ;;
     "Quit")
-      break
       exit
       ;;
-    *) echo "invalid option $REPLY" ;;
+    *) cecho "RED" "invalid option $REPLY" ;;
     esac
   done
 elif [ -f /opt/procursus/bin/apt ]; then
-  echo "Procursus found."
+  cecho "GREEN" "[!] Procursus is installed"
   pkg="sudo apt"
 elif command -v brew >/dev/null; then
-  echo "[!] Homebrew found!"
+  cecho "GREEN" "[!] Homebrew is installed!"
   pkg="brew"
 else
-  echo "[!] Procursus nor Homebrew were found."
+  cecho "YELLOW" "[!] Procursus nor Homebrew were found."
   if [ "$verscheck" -eq 1 ]; then
-    echo "[!] Procursus is not compatible with your macOS version."
+    cecho "RED" "[!] Procursus is not compatible with your macOS version."
   else
     echo "Would you like to install Procursus?"
     read -p "[y/n]" installpro
@@ -65,41 +77,41 @@ else
       exec ./procursus-install-macOS.sh
     fi
   fi
-  echo "Homebrew not found. Install instructions can be found at https://brew.sh"
+  cecho "YELLOW" "Homebrew not found. Install instructions can be found at https://brew.sh"
   exit 3
 fi
 
-if [ ! "$(command -v futurerestore)" ] && [ -z "$HOME/FutureRestoreGUI/.extracted/futurerestore" ]; then
-  echo "futurerestore not found. Download at https://github.com/futurerestore/futurerestore"
+if [ ! "$(command -v futurerestore)" ] && [ ! -e "$HOME/FutureRestoreGUI/extracted/futurerestore" ]; then
+  cecho "RED" "futurerestore not found. Download at https://github.com/futurerestore/futurerestore"
   sleep 5
   open https://github.com/futurerestore/futurerestore
   exit 3
 elif command -v futurerestore; then
-  echo "[!] futurerestore found."
-elif [ -z "$HOME/FutureRestoreGUI/.extracted/futurerestore" ]; then
-  echo "[!] Located futurerestore downloaded by FutureRestoreGUI."
+  cecho "GREEN" "[!] futurerestore is installed!"
+elif [ -e "$HOME/FutureRestoreGUI/extracted/futurerestore" ]; then
+  cecho "GREEN" "[!] Located futurerestore downloaded by FutureRestoreGUI."
 fi
 
 if command -v irecovery >/dev/null; then
-  echo "[!] irecovery found."
+  cecho "GREEN" "[!] irecovery is installed!"
 else
-  echo "[!] irecovery not found. Installing..."
+  cecho "YELLOW" "[!] irecovery not found. Installing..."
   if [ "$pkg" = "sudo apt" ]; then
     $pkg install libirecovery-utils
   elif [ "$pkg" = "brew" ]; then
-    brew install libirecovery
+    $pkg install libirecovery
   fi
 fi
 
 if command -v git > /dev/null; then
-  echo "[!] git found!"
+  cecho "GREEN" "[!] git is installed!"
 else
-  echo "[!] git not found. Installing..."
+  cecho "YELLOW" "[!] git not found. Installing..."
   $pkg install git
 fi
 
-if command -v clang; then
-  echo "[!] clang found!"
+if command -v clang > /dev/null; then
+  cecho "GREEN" "[!] clang is installed!"
 else
     if [ "$pkg" = "sudo apt" ]; then
     $pkg install clang
@@ -108,25 +120,29 @@ else
   fi
 fi
 
-if [[ $(python3 --version) != *"Python 3."* ]]; then
-  echo "python3 not found. Installing..."
+if ! command -v python3 > /dev/null; then
+  cecho "YELLOW" "python3 not found. Installing..."
   $pkg install python3
+  else 
+  cecho "GREEN" "[!] python3 is installed!"
 fi
 
 if python3 -m pip | grep "No module named pip"; then
-  echo "pip not found. installing"
+  cecho "YELLOW" "[!] pip not found. installing"
   python3 -m ensurepip
 else
-  echo "[!] pip found."
+  cecho "GREEN" "[!] pip is installed!."
 fi
 
-if [ $(python3 -m pip list | grep -c "pyimg4") == 0 ]; then
-  echo "pyimg4 not found. Installing..."
+if [ "$(python3 -m pip list | grep -c "pyimg4")" == 0 ]; then
+  cecho "YELLOW" "[!] pyimg4 not found. Installing..."
   python3 -m pip install pyimg4 || error_exit "[!] pyimg4 failed to install"
+  else 
+  cecho "GREEN" "[!] pyimg4 is installed!" 
 fi
 
 if [ ! -e "/usr/local/bin/img4" ]; then
-  echo "img4 not found. Downloading..."
+  cecho "YELLOW" "[!] img4 not found. Downloading..."
   curl --progress-bar -o img4lib.tar.gz -L https://github.com/xerub/img4lib/releases/download/1.0/img4lib-2020-10-27.tar.gz || error_exit "Download failed."
   tar -xvf img4lib.tar.gz
   sudo mkdir -p /usr/local/bin
@@ -139,8 +155,8 @@ if [ ! -e "/usr/local/bin/img4" ]; then
 fi
 
 if [ ! -e "/usr/local/bin/img4tool" ]; then
-  echo "img4tool not found. Downloading..."
-  curl --progress-bar -OL https://github.com/tihmstar/img4tool/releases/download/197/buildroot_macos-latest.zip
+  cecho "YELLOW" "[!] img4tool not found. Downloading..."
+  curl --progress-bar -OL https://github.com/tihmstar/img4tool/releases/download/197/buildroot_macos-latest.zip || error_exit "Download failed."
   unzip buildroot_macos-latest.zip
   usr_local=buildroot_macos-latest/usr/local
   mv -v $usr_local/bin/img4tool /usr/local/bin/
@@ -160,7 +176,7 @@ fi
 cd bin
 
 if [ ! -e "./gaster" ]; then
-  echo "gaster not found. Downloading..."
+  cecho "YELLOW" "[!] gaster not found. Downloading..."
   git clone https://github.com/0x7ff/gaster.git gaster_git
   cd gaster_git
   make
@@ -169,20 +185,24 @@ if [ ! -e "./gaster" ]; then
   rm -rf gaster_git/
   chmod 755 gaster
   xattr -d com.apple.quarantine gaster
+  else
+  cecho "GREEN" "[!] gaster found!"
 fi
 
 if [ ! -e "./iBoot64Patcher" ]; then
-  echo "iBoot64Patcher not found. Downloading..."
+  cecho "YELLOW" "[!] iBoot64Patcher not found. Downloading..."
   curl --progress-bar -OL https://nightly.link/Arna13/iBoot64Patcher/actions/runs/3176527177/iBoot64Patcher-${OS}-RELEASE.zip
   unzip iBoot64Patcher-${OS}-RELEASE.zip
   tar -xvf iBoot64Patcher-${OS}-*-RELEASE.tar.xz
   rm -rf iBoot64Patcher-*
   chmod 755 iBoot64Patcher
   xattr -d com.apple.quarantine iBoot64Patcher
+   else
+  cecho "GREEN" "[!] iBoot64Patcher found!"
 fi
 
 if [ ! -e "./Kernel64Patcher" ]; then
-  echo "Kernel64Patcher not found. Downloading..."
+  cecho "YELLOW" "[!] Kernel64Patcher not found. Downloading..."
   git clone https://github.com/iSuns9/Kernel64Patcher.git Kernel64Patcher_git
   cd Kernel64Patcher_git
   clang Kernel64Patcher.c -o Kernel64Patcher
@@ -191,10 +211,12 @@ if [ ! -e "./Kernel64Patcher" ]; then
   rm -rf Kernel64Patcher_git/
   chmod 755 Kernel64Patcher
   xattr -d com.apple.quarantine Kernel64Patcher
+   else
+  cecho "GREEN" "[!] Kernel64Patcher found!"
 fi
 
 if [ ! -e "./asr64_patcher" ]; then
-  echo "asr64_patcher not found. Downloading..."
+  cecho "YELLOW" "[!] asr64_patcher not found. Downloading..."
   git clone https://github.com/exploit3dguy/asr64_patcher.git asr64_patcher_git
   cd asr64_patcher_git
   clang asr64_patcher.c -o asr64_patcher
@@ -203,10 +225,12 @@ if [ ! -e "./asr64_patcher" ]; then
   rm -rf asr64_patcher_git/
   chmod 755 asr64_patcher
   xattr -d com.apple.quarantine asr64_patcher
+   else
+  cecho "GREEN" "[!] asr64_patcher found!"
 fi
 
 if [ ! -e "./restored_external64_patcher" ]; then
-  echo "[!] restored_external64_patcher not found. Downloading..."
+  cecho "YELLOW" "[!] restored_external64_patcher not found. Downloading..."
   git clone https://github.com/iSuns9/restored_external64patcher.git restored_external64patcher_git
   cd restored_external64patcher_git
   clang restored_external64_patcher.c -o restored_external64_patcher
@@ -215,28 +239,41 @@ if [ ! -e "./restored_external64_patcher" ]; then
   rm -rf restored_external64patcher_git/
   chmod 755 restored_external64_patcher
   xattr -d com.apple.quarantine restored_external64_patcher
+   else
+  cecho "GREEN" "[!] restored_external64_patcher found!"
 fi
 
-if [ ! -e "./ldid2" ]; then
-  echo "ldid2 not found. Downloading..."
+if [ ! -e "./ldid2" ] && [ "$(command -v ldid)" != "/opt/procursus/bin/ldid" ]; then
+  cecho "YELLOW" "ldid2 not found. Downloading..."
   if [ "$OS" = "macOS-x86_64" ]; then 
   OS=macos_x86_64
   elif [ "$OS" = "macOS-arm64" ]; then
   OS=macos_arm64
   fi
+
+  if [ "$pkg" = "sudo apt" ]; then
+  $pkg install ldid
+  else
   curl --progress-bar -o ldid2 -L https://github.com/ProcursusTeam/ldid/releases/download/v2.1.5-procursus5/ldid_${OS} || error_exit "Download failed."
   chmod 755 ldid2
   xattr -d com.apple.quarantine ldid2
+  fi
+   else
+  cecho "GREEN" "[!] ldid2 found!"
 fi
 
 if [ ! -e "./tsschecker" ]; then
-  echo "tsschecker not found. Downloading..."
+  cecho "YELLOW" "tsschecker not found. Downloading..."
   curl --progress-bar -o tsschecker.zip -L https://github.com/tihmstar/tsschecker/releases/download/304/tsschecker_macOS_v304.zip || error_exit "Download failed."
   unzip tsschecker.zip
   rm tsschecker.zip
   chmod 755 tsschecker
   xattr -d com.apple.quarantine tsschecker
+   else
+  cecho "GREEN" "[!] tsschecker found!"
 fi
 
 cd ../
+
+cecho "GREEN" "### Dependency installation finished ###"
 touch .requirements_done
